fix: resolved tests errors
